replace token for enrollment with util function
